fixed the half day salary calculation feature

+++ JayathilakaPosBackedWithPrinter/src/main/java/com/jayathilakaRiceProducts/JayathilakaPosBackedWithPrinter/dto/EmployeeDTO.java
@@ -19,15 +19,15 @@ public class EmployeeDTO {
     private double loan_to_pay ;  // remain amount for pay ;ගෙවිය යුතු මුළු ණය මුදල
     private double loan_payment_for_month ; //පෙර හිග මුදල්
     private double calculated_salary;
-    private int worked_days_count; //මුළු වැඩ කරන දින ගණන; how many day employee has worked
-    private int should_work_dates_total; //වැඩ කරන ලද මුළු දිනයන්; how many day employee should work
-    private int extraWorkedDays;
+    private double worked_days_count; //මුළු වැඩ කරන දින ගණන; how many day employee has worked
+    private double should_work_dates_total; //වැඩ කරන ලද මුළු දිනයන්; how many day employee should work
+    private double extraWorkedDays;
     private int month;
 
     public EmployeeDTO() {
     }
 
-    public EmployeeDTO(String name, String position, double monthlySalary, double ETF, double otherAllowances, double specialSupports, double advance_payments, double other_deductions, double loan_to_pay, double loan_payment_for_month, double calculated_salary, int worked_days_count, int should_work_dates_total, int extraWorkedDays, int month) {
+    public EmployeeDTO(String name, String position, double monthlySalary, double ETF, double otherAllowances, double specialSupports, double advance_payments, double other_deductions, double loan_to_pay, double loan_payment_for_month, double calculated_salary, double worked_days_count, double should_work_dates_total, double extraWorkedDays, int month) {
         this.name = name;
         this.position = position;
         this.monthlySalary = monthlySalary;
